requirements.txt
fastapi[all]==0.95.0
uvicorn[standard]==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.16
grpcio==1.56.2
grpcio-tools==1.56.2
